+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - bugfix the live search in the dropdown for selecting groups for job actions from type `add_group` or `remove_group`;
 - now `condition_on_execute` is properly used for jobs form type `tasks`
 - add margin to the home button on missing page
+- #357 - is_multiple in select does not save data properly.
 
 # v6.2.0
 ### New features